+++ src/easyscience/Fitting/minimizers/__init__.py
@@ -7,8 +7,6 @@ __version__ = '0.1.0'
 
 import warnings
 
-from .fitting_template import FitError  # noqa: F401, E402
-
 imported = -1
 try:
     from .lmfit import lmfit  # noqa: F401, E402
@@ -25,13 +23,16 @@ except ImportError:
     # TODO make this a proper message (use logging?)
     warnings.warn('bumps has not been installed.', ImportWarning, stacklevel=2)
 try:
-    from .DFO_LS import DFO  # noqa: F401, E402
+    from .dfo import DFO  # noqa: F401, E402
 
     imported += 1
 except ImportError:
     # TODO make this a proper message (use logging?)
     warnings.warn('dfo-ls has not been installed.', ImportWarning, stacklevel=2)
 
-from .fitting_template import FittingTemplate  # noqa: E402
+from .fitting_base import FittingBase  # noqa: E402
 
-engines: list = FittingTemplate._engines
+engines: list = FittingBase._engines
+
+from .utils import FitError  # noqa: F401, E402
+from .utils import FitResults  # noqa: F401, E402
