+++ src/easyscience/Fitting/minimizers/fitting_base.py
@@ -1,7 +1,6 @@
 __author__ = 'github.com/wardsimon'
 __version__ = '0.1.0'
 
-
 #  SPDX-FileCopyrightText: 2023 EasyScience contributors  <core@easyscience.software>
 #  SPDX-License-Identifier: BSD-3-Clause
 #  © 2021-2023 Contributors to the EasyScience project <https://github.com/easyScience/EasyScience
@@ -15,8 +14,10 @@ from typing import Union
 
 import numpy as np
 
+from .utils import FitResults
 
-class FittingTemplate(metaclass=ABCMeta):
+
+class FittingBase(metaclass=ABCMeta):
     """
     This template class is the basis for all fitting engines in `EasyScience`.
     """
@@ -194,79 +195,79 @@ class FittingTemplate(metaclass=ABCMeta):
         return error_matrix
 
 
-class FitResults:
-    """
-    At the moment this is just a dummy way of unifying the returned fit parameters.
-    """
+# class FitResults:
+#     """
+#     At the moment this is just a dummy way of unifying the returned fit parameters.
+#     """
 
-    __slots__ = [
-        'success',
-        'fitting_engine',
-        'fit_args',
-        'p',
-        'p0',
-        'x',
-        'x_matrices',
-        'y_obs',
-        'y_calc',
-        'y_err',
-        'engine_result',
-        'total_results',
-    ]
+#     __slots__ = [
+#         'success',
+#         'fitting_engine',
+#         'fit_args',
+#         'p',
+#         'p0',
+#         'x',
+#         'x_matrices',
+#         'y_obs',
+#         'y_calc',
+#         'y_err',
+#         'engine_result',
+#         'total_results',
+#     ]
 
-    def __init__(self):
-        self.success = False
-        self.fitting_engine = None
-        self.fit_args = {}
-        self.p = {}
-        self.p0 = {}
-        self.x = np.ndarray([])
-        self.x_matrices = np.ndarray([])
-        self.y_obs = np.ndarray([])
-        self.y_calc = np.ndarray([])
-        self.y_err = np.ndarray([])
-        self.engine_result = None
-        self.total_results = None
+#     def __init__(self):
+#         self.success = False
+#         self.fitting_engine = None
+#         self.fit_args = {}
+#         self.p = {}
+#         self.p0 = {}
+#         self.x = np.ndarray([])
+#         self.x_matrices = np.ndarray([])
+#         self.y_obs = np.ndarray([])
+#         self.y_calc = np.ndarray([])
+#         self.y_err = np.ndarray([])
+#         self.engine_result = None
+#         self.total_results = None
 
-    @property
-    def n_pars(self):
-        return len(self.p)
+#     @property
+#     def n_pars(self):
+#         return len(self.p)
 
-    @property
-    def residual(self):
-        return self.y_obs - self.y_calc
+#     @property
+#     def residual(self):
+#         return self.y_obs - self.y_calc
 
-    @property
-    def chi2(self):
-        return ((self.residual / self.y_err) ** 2).sum()
+#     @property
+#     def chi2(self):
+#         return ((self.residual / self.y_err) ** 2).sum()
 
-    @property
-    def reduced_chi(self):
-        return self.chi2 / (len(self.x) - self.n_pars)
-
-
-class NameConverter:
-    def __init__(self):
-        from easyscience import borg
-
-        self._borg = borg
-
-    def get_name_from_key(self, item_key: int) -> str:
-        return getattr(self._borg.map.get_item_by_key(item_key), 'name', '')
-
-    def get_item_from_key(self, item_key: int) -> object:
-        return self._borg.map.get_item_by_key(item_key)
-
-    def get_key(self, item: object) -> int:
-        return self._borg.map.convert_id_to_key(item)
+#     @property
+#     def reduced_chi(self):
+#         return self.chi2 / (len(self.x) - self.n_pars)
 
 
-class FitError(Exception):
-    def __init__(self, e: Exception = None):
-        self.e = e
+# class NameConverter:
+#     def __init__(self):
+#         from easyscience import borg
 
-    def __str__(self) -> str:
-        s = ''
-        if self.e is not None:
-            s = f'{self.e}\n'
-        return s + 'Something has gone wrong with the fit'
+#         self._borg = borg
+
+#     def get_name_from_key(self, item_key: int) -> str:
+#         return getattr(self._borg.map.get_item_by_key(item_key), 'name', '')
+
+#     def get_item_from_key(self, item_key: int) -> object:
+#         return self._borg.map.get_item_by_key(item_key)
+
+#     def get_key(self, item: object) -> int:
+#         return self._borg.map.convert_id_to_key(item)
+
+
+# class FitError(Exception):
+#     def __init__(self, e: Exception = None):
+#         self.e = e
+
+#     def __str__(self) -> str:
+#         s = ''
+#         if self.e is not None:
+#             s = f'{self.e}\n'
+#         return s + 'Something has gone wrong with the fit'
